reorganised helper functions on the model of the language
