+++ packages/meta/src/utils/index.ts
@@ -1,8 +1,10 @@
 export * from "./Checker";
 export * from "./Helpers";
-export * from "./ModelHelpers";
+export * from "./MetaLogger";
+export * from "./GenerationHelpers";
 export * from "./Names";
 export * from "./PathProvider";
 export * from "./PiDefinitionElement";
+export * from "./LangUtil";
 export * from "./PiParser";
 export * from "./Roles";
